created largeimage and app tests
