+++ src/components/LargeImage.jsx
@@ -2,9 +2,7 @@ import React from 'react';
 
 const LargeImage = ({ imageUrl }) => {
   return (
-    <div >
-      <img src={imageUrl} className="largeImage" width="560" height="300" />
-    </div>
+    <img src={imageUrl} className="largeImage" width="560" height="300" />
   )
 }
